Add files via upload
CREDIT MANAGEMENT SYSTEM (for educational purposes)

The system will have three different users with access to different screens.
We begin with the general user registration. We have a put a validation number only on the phone number for 10 digits and password to be a combination of 8 alphanumeric characters.
The general user can login with the registered username and password combination. In case the password is lost, we can reset with the correct combination of username and date of birth registered earlier using the password manager.
In case the user is authenticated, the credit management system landing page is shown. This page has separate logins for ADMIN, Students, and teachers. Here we will require to register for the specific role as admin, student or as teacher. 
If the user registers as an admin he can go and create student profiles with contact details and class. Validation has been placed on uniqueness of the profile id, phone number for 10 numeric and class for between 9 and 12. The admin can search, add, select, update, delete and reset all the entries using the button on the side. The data can be seen in the table.
If the user registers as a student, the user can now login and enter name, age and select the class to go to the credit booster questionnaire with 11 questions with logical choices. A yes gives some points and a No gives 0 points. On Submitting the questionnaire , the credits are displayed to the student and successfully submitted . If you need to change the submission  , the specific answer has to be corrected and then submitted.
If the user registers as a teacher, on successful login , he/she just needs to enter her name and select the class , the name age class credit will be displayed as a report on the screen. If the teacher wants the information offline, she will press save. The data will be saved on the desktop for further reference .